move fill,stroke code to GridLayouter
now grid fill and stroke works just like table's
